- Files in Doc: Updates for the website - Modules/objc/pyobjc.h : Version number updated - The other changes are preliminary 10.1 support. If you disable FFI this   compiles, but will cause a crash when actually using the objc module.

+++ pyobjc/Scripts/CodeGenerators/enum_generator.py
@@ -73,6 +73,8 @@ def process_file(outfp, filename):
 
 
 def generate(dirname, fn = None):
+	if not os.path.exists(dirname): return
+
 	if fn:
 		fp = open(fn, 'w')
 	else:
